backend/api/requirements/common.txt
datacube==1.8.12
Flask==2.1.1
Flask-Cors==3.0.8
Flask-HTTPAuth==4.1.0
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.25.0
GeoAlchemy2==0.11.1
geopandas==0.12.2
influxdb-client[ciso]==1.28.0
itsdangerous==2.0.1
marshmallow-sqlalchemy==0.24.2
mikeio==1.3.1
openpyxl==3.0.10
OWSLib==0.28.1
pandas==1.4.2
prophet==1.1.1
psycopg2-binary==2.9.1
pyodbc==4.0.34
pyproj==3.4.1
python-cas==1.6.0
python-dotenv==0.14.0
pywr==1.18.0
rasterio==1.3.6
rioxarray==0.13.4
Rtree==1.0.1
scikit-learn==1.2.2
shapely==2.0.1
tensorflow==2.9.1
Werkzeug==2.0.0